Se cambió el nombre de Pregunta a QuizGame
